refactor(cli): consolidate transcript finalize-status predicate
`subscribeStreamLog.shouldFinalizeDeferredEntries` and
`subscribeStreamStatus.FINAL_TRANSCRIPT_STATUSES` independently
enumerated the same four stream statuses (ERROR/READY/STOPPED/WAITING)
that promote deferred entries into `<Static>` scrollback. Promote the
predicate to `transcript.ts` next to `finalizeAssistantTranscriptEntries`
and let both subscribers call it, so the rule lives in one place.
