+++ packages/cli/src/chat/tui/state/subscribeStreamStatus.ts
@@ -1,23 +1,18 @@
 // Mirror `StreamStatusService.onDidChange` into the per-stream status signal.
 
 import { StreamStatusService } from '@agent/runtime/StreamStatusService';
-import { STREAM_STATUS, type StreamStatus } from '@shared/schemas';
 
 import { patchStream } from './cliState';
 import { syncStreamLog } from './subscribeStreamLog';
-import { finalizeAssistantTranscriptEntries } from './transcript';
-
-const FINAL_TRANSCRIPT_STATUSES: ReadonlySet<StreamStatus> = new Set([
-  STREAM_STATUS.ERROR,
-  STREAM_STATUS.READY,
-  STREAM_STATUS.STOPPED,
-  STREAM_STATUS.WAITING,
-]);
+import {
+  finalizeAssistantTranscriptEntries,
+  isFinalTranscriptStatus,
+} from './transcript';
 
 export function subscribeStreamStatus(): () => void {
   return StreamStatusService.onDidChange((change) => {
     syncStreamLog(change.streamId);
-    if (FINAL_TRANSCRIPT_STATUSES.has(change.status)) {
+    if (isFinalTranscriptStatus(change.status)) {
       finalizeAssistantTranscriptEntries(change.streamId);
     }
     patchStream(change.streamId, (slice) => ({
